% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadsum815-package.R
\docType{package}
\name{quadsum815-package}
\alias{quadsum815}
\alias{quadsum815-package}
\title{quadsum815: Various R and Rcpp Implementations of a Quadratic Sum}
\description{
This package is in fulfillment of the requirements of Homework 00 for BIOSTAT 815. quadsum815 provides 
    five different implementations of functions to compute a single scalar value of the quadratic sum x'Ay. The functions 
    are as follows: (1) quadsumR: A loop-based R implementation, (2) quadsumS: A loop-free R implementation, (3) quadsumT: 
    A matrix-based R implementation, (4) quadsumC: A loop-based Rcpp implementation, and (5) quadsumB: A blockwise 
    loop-based Rcpp implementation.
}
\keyword{internal}
