% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadsumB}
\alias{quadsumB}
\title{A blockwise loop-based Rcpp implementation for computing \eqn{x'Ay}}
\usage{
quadsumB(x, y, A, blkSize)
}
\arguments{
\item{x}{A size \eqn{n} vector}

\item{y}{A size \eqn{m} vector}

\item{A}{A size (\eqn{n x m}) matrix}

\item{blkSize}{The size of blocks (in #rows or #cols) in calculating \eqn{x'Ay}}
}
\value{
A scalar value evaluating \eqn{x'Ay}
}
\description{
This function computes \eqn{x'Ay} using blockwise nested loop using Rcpp
}
\examples{
quadsumB(rep(1,100),rep(1,100),matrix(1,100,100), 100)  # [1] 10000
}
