% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadsumS.r
\name{quadsumS}
\alias{quadsumS}
\title{A loop-free R implementation for computing \eqn{x'Ay}}
\usage{
quadsumS(x, y, A)
}
\arguments{
\item{x}{A size \eqn{n} vector}

\item{y}{A size \eqn{m} vector}

\item{A}{A size (\eqn{n x m}) matrix}
}
\value{
A scalar value evaluating \eqn{x'Ay}
}
\description{
A loop-free R implementation for computing \eqn{x'Ay}
}
\examples{
quadsumS(rep(1,100),rep(1,100),matrix(1,100,100))  # [1] 10000
}
