% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadsumT.r
\name{quadsumT}
\alias{quadsumT}
\title{Matrix-based computation of x'Ay}
\usage{
quadsumT(x, y, A)
}
\arguments{
\item{x}{A size n vector}

\item{y}{A size m vector}

\item{A}{A size (n x m) matrix}
}
\value{
A scalar value evaluating x'Ay
}
\description{
Matrix-based computation of x'Ay
}
