\name{quadsum815-package}
\alias{quadsum815-package}
\alias{quadsum815}
\docType{package}
\title{
  Various implementations of quadratic sum  
}
\description{
  This contains various implementations of quadratic sum covered in BIOSTAT815  
}
\author{
Hyun Min Kang

Maintainer: Hyun Min Kang <hmkang@umich.edu>
}
\keyword{ package }
