% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadsumB}
\alias{quadsumB}
\title{A blockwise loop implementation for computing x'Ay using Rcpp}
\usage{
quadsumB(x, y, A, blkSize)
}
\arguments{
\item{x}{A size n vector}

\item{y}{A size m vector}

\item{A}{A size (n x m) matrix}

\item{blkSize}{The size of blocks (in #rows or #cols) in calculating x'Ay}
}
\value{
A scalar value evaluating x'Ay
}
\description{
This function computes x'Ay using blockwise nested loop using Rcpp
}
