% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadsumC}
\alias{quadsumC}
\title{A loop-based Rcpp implementation for computing x'Ay}
\usage{
quadsumC(x, y, A)
}
\arguments{
\item{x}{A size n vector}

\item{y}{A size m vector}

\item{A}{A size (n x m) matrix}
}
\value{
A scalar value evaluating x'Ay
}
\description{
A loop-based Rcpp implementation for computing x'Ay
}
