% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadsumR.r
\name{quadsumR}
\alias{quadsumR}
\title{A loop-based R implementation for computing x'Ay}
\usage{
quadsumR(x, y, A)
}
\arguments{
\item{x}{A size n vector}

\item{y}{A size m vector}

\item{A}{A size (n x m) matrix}
}
\value{
A scalar value evaluating x'Ay
}
\description{
A loop-based R implementation for computing x'Ay
}
