% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadsumS.r
\name{quadsumS}
\alias{quadsumS}
\title{A loop-free R implementation for computing x'Ay}
\usage{
quadsumS(x, y, A)
}
\arguments{
\item{x}{A size n vector}

\item{y}{A size m vector}

\item{A}{A size (n x m) matrix}
}
\value{
A scalar value evaluating x'Ay
}
\description{
A loop-free R implementation for computing x'Ay
}
